Move multer-config & auth files to a middlewares directory
